https://online.sbis.ru/opendoc.html?guid=5b04c8f6-ab31-401d-9860-3ccedf638157 Разделил демки на разные файлы
