--- Controls-demo/treeGrid/ItemPadding/All/All.wml
@@ -1,122 +1,0 @@
-<div class="controlsDemo__wrapper">
-
-
-    <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-default">
-        <div class="controls-text-label_theme-{{_options.theme}}">Стандартные отступы</div>
-        <Controls.treeGrid:View
-                keyProperty="id"
-                source="{{_viewSource}}"
-                columns="{{_columns}}"
-                parentProperty="parent"
-                nodeProperty="type"/>
-    </div>
-
-    <div class="controlsDemo__cell demo-ItemPaddingAll__vertical">
-        <div class="controls-text-label_theme-{{_options.theme}}">Вертикальные отступы</div>
-        <div class="controlsDemo__wrapper controlsDemo__flexRow">
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-vertical-m">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру M</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding top="null" bottom="null"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-vertical-l">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру L</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding top="L" bottom="L"/>
-                </Controls.treeGrid:View>
-            </div>
-        </div>
-    </div>
-
-    <div class="controlsDemo__cell demo-ItemPaddingAll__horizontal">
-        <div class="controls-text-label_theme-{{_options.theme}}">Горизонтальные отступы</div>
-        <div class="controlsDemo__wrapper controlsDemo__flexRow">
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-null">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру null</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="null" right="null"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-s">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру S</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="s" right="s"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-m">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру M</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="m" right="m"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-l">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру L</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="L" right="L"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-xl">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру XL</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="XL" right="XL"/>
-                </Controls.treeGrid:View>
-            </div>
-
-            <div class="controlsDemo__cell controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-xxl">
-                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру XXL</div>
-                <Controls.treeGrid:View
-                        keyProperty="id"
-                        source="{{_viewSource}}"
-                        columns="{{_columns}}"
-                        parentProperty="parent"
-                        nodeProperty="type">
-                    <ws:itemPadding left="XXL" right="XXL"/>
-                </Controls.treeGrid:View>
-            </div>
-        </div>
-    </div>
-
-</div>
