--- Controls-demo/treeGrid/ItemPadding/All/Index.ts
@@ -1,19 +1,0 @@
-import {Control, TemplateFunction} from "UI/Base"
-import * as Template from "wml!Controls-demo/treeGrid/ItemPadding/All/All"
-import {Memory} from "Types/source"
-import {Gadgets} from "../../DemoHelpers/DataCatalog"
-
-import 'css!Controls-demo/Controls-demo'
-
-export default class extends Control {
-    protected _template: TemplateFunction = Template;
-    protected _viewSource: Memory;
-    protected _columns = Gadgets.getColumnsForFlat();
-
-    protected _beforeMount() {
-        this._viewSource = new Memory({
-            keyProperty: 'id',
-            data: Gadgets.getFlatData()
-        });
-    }
-}
