+++ Controls-demo/treeGrid/ItemPadding/BaseIndex.ts
@@ -1,0 +1,24 @@
+import {Control, TemplateFunction, IControlOptions} from 'UI/Base';
+import {Memory} from 'Types/source';
+import 'css!Controls-demo/Controls-demo';
+
+import {Gadgets} from '../DemoHelpers/DataCatalog';
+
+interface BaseIndexOptions extends IControlOptions {
+    showTitle: boolean;
+}
+
+export default class extends Control<BaseIndexOptions> {
+    protected _template: TemplateFunction;
+    protected _viewSource: Memory;
+    protected _columns = Gadgets.getColumnsForFlat();
+    protected _showTitle: boolean;
+
+    protected _beforeMount(options?: BaseIndexOptions, contexts?: object, receivedState?: void): Promise<void> | void {
+        this._showTitle = !options || options.showTitle !== false;;
+        this._viewSource = new Memory({
+            keyProperty: 'id',
+            data: Gadgets.getFlatData()
+        });
+    }
+}
