+++ Controls-demo/treeGrid/ItemPadding/Custom/Custom.wml
@@ -1,10 +1,13 @@
-<div class="controlsDemo__wrapper controlsDemo_fixedWidth300 controlsDemo_fixedWidth300 controlsDemo_treeGrid-iP-custom">
-    <Controls.treeGrid:View
-            keyProperty="id"
-            source="{{_viewSource}}"
-            columns="{{_columns}}"
-            parentProperty="parent"
-            nodeProperty="type">
-        <ws:itemPadding left="XXL" right="null" />
-    </Controls.treeGrid:View>
+<div class="controlsDemo__wrapper ws-flexbox ws-flex-wrap">
+    <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo_treeGrid-iP-custom">
+        <div if="{{_showTitle}}" class="controls-text-label_theme-{{_options.theme}}">Настраиваемые отступы</div>
+        <Controls.treeGrid:View
+                keyProperty="id"
+                source="{{_viewSource}}"
+                columns="{{_columns}}"
+                parentProperty="parent"
+                nodeProperty="type">
+            <ws:itemPadding left="XXL" right="null"/>
+        </Controls.treeGrid:View>
+    </div>
 </div>
