+++ Controls-demo/treeGrid/ItemPadding/Horizontal/Horizontal.wml
@@ -1,0 +1,80 @@
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__cell demo-ItemPaddingAll__horizontal">
+        <div if="{{_showTitle}}" class="controls-text-label_theme-{{_options.theme}}">Горизонтальные отступы</div>
+        <div class="controlsDemo__wrapper ws-flexbox ws-flex-wrap">
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-null">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер null</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="null" right="null"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-s">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер S</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="s" right="s"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-m">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер M</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="m" right="m"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-l">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер L</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="L" right="L"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell  controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-xl">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер XL</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="XL" right="XL"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell  controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-horizontal-xxl">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы размер XXL</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding left="XXL" right="XXL"/>
+                </Controls.treeGrid:View>
+            </div>
+        </div>
+    </div>
+
+</div>
