+++ Controls-demo/treeGrid/ItemPadding/Horizontal/Index.ts
@@ -1,7 +1,7 @@
 import {TemplateFunction} from 'UI/Base';
 import Baseindex from '../BaseIndex';
 
-import * as Template from 'wml!Controls-demo/treeGrid/ItemPadding/Custom/Custom';
+import * as Template from 'wml!Controls-demo/treeGrid/ItemPadding/Horizontal/Horizontal';
 
 export default class extends Baseindex {
     protected _template: TemplateFunction = Template;
