+++ Controls-demo/treeGrid/ItemPadding/ItemPadding.wml
@@ -4,19 +4,36 @@
         <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
                                  state="/Controls-demo/app/:app"
                                  app="{{ 'Controls-demo/treeGrid/ItemPadding/Custom/Index' }}">
-            <a href="{{content.href}}">Настройка отступов</a>
+            <a href="{{content.href}}">Настраиваемые отступы</a>
         </Router.router:Reference>
-        <Controls-demo.treeGrid.ItemPadding.Custom.Index/>
+        <Controls-demo.treeGrid.ItemPadding.Custom.Index showTitle="{{false}}"/>
     </div>
-
 
     <div class="controlsDemo__cell">
         <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
                                  state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/treeGrid/ItemPadding/All/Index' }}">
-            <a href="{{content.href}}">Все размеры отступов</a>
+                                 app="{{ 'Controls-demo/treeGrid/ItemPadding/Standard/Index' }}">
+            <a href="{{content.href}}">Стандартные отступы</a>
         </Router.router:Reference>
-        <Controls-demo.treeGrid.ItemPadding.All.Index/>
+        <Controls-demo.treeGrid.ItemPadding.Standard.Index showTitle="{{false}}"/>
+    </div>
+
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/treeGrid/ItemPadding/Horizontal/Index' }}">
+            <a href="{{content.href}}">Горизонтальные отступы</a>
+        </Router.router:Reference>
+        <Controls-demo.treeGrid.ItemPadding.Horizontal.Index showTitle="{{false}}"/>
+    </div>
+
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/treeGrid/ItemPadding/Vertical/Index' }}">
+            <a href="{{content.href}}">Вертикальные отступы</a>
+        </Router.router:Reference>
+        <Controls-demo.treeGrid.ItemPadding.Vertical.Index showTitle="{{false}}"/>
     </div>
 
 </div>
