+++ Controls-demo/treeGrid/ItemPadding/Standard/Standard.wml
@@ -1,0 +1,11 @@
+<div class="controlsDemo__wrapper ws-flexbox ws-flex-wrap">
+    <div class="controlsDemo__cell  controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-default">
+        <div if="{{_showTitle}}" class="controls-text-label_theme-{{_options.theme}}">Стандартные отступы</div>
+        <Controls.treeGrid:View
+                keyProperty="id"
+                source="{{_viewSource}}"
+                columns="{{_columns}}"
+                parentProperty="parent"
+                nodeProperty="type"/>
+    </div>
+</div>
