+++ Controls-demo/treeGrid/ItemPadding/Vertical/Index.ts
@@ -1,7 +1,7 @@
 import {TemplateFunction} from 'UI/Base';
 import Baseindex from '../BaseIndex';
 
-import * as Template from 'wml!Controls-demo/treeGrid/ItemPadding/Custom/Custom';
+import * as Template from 'wml!Controls-demo/treeGrid/ItemPadding/Vertical/Vertical';
 
 export default class extends Baseindex {
     protected _template: TemplateFunction = Template;
