+++ Controls-demo/treeGrid/ItemPadding/Vertical/Vertical.wml
@@ -1,0 +1,31 @@
+<div class="controlsDemo__wrapper">
+    <div class="controlsDemo__cell demo-ItemPaddingAll__vertical">
+        <div if="{{_showTitle}}" class="controls-text-label_theme-{{_options.theme}}">Вертикальные отступы</div>
+        <div class="controlsDemo__wrapper ws-flexbox ws-flex-wrap">
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-vertical-m">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру M</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding top="null" bottom="null"/>
+                </Controls.treeGrid:View>
+            </div>
+
+            <div class="controlsDemo__cell controlsDemo__width-s_theme-{{_options.theme}} controlsDemo__mr2 controlsDemo_treeGrid-iP-all-vertical-l">
+                <div class="controls-text-label_theme-{{_options.theme}}">Отступы соответствующие размеру L</div>
+                <Controls.treeGrid:View
+                        keyProperty="id"
+                        source="{{_viewSource}}"
+                        columns="{{_columns}}"
+                        parentProperty="parent"
+                        nodeProperty="type">
+                    <ws:itemPadding top="L" bottom="L"/>
+                </Controls.treeGrid:View>
+            </div>
+        </div>
+    </div>
+</div>
